@@ -2,7 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {Store} from '@ngrx/store';
 import {AppState} from '../../app.state';
 import {Game} from '../shared/game';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import * as gameActions from '../store/games.actions';
 import {getAllGames} from '../store/games.reducers';
 import {PlatformsService} from '../shared/platforms.service';
@@ -32,7 +32,6 @@ export class GameListComponent implements OnInit {
 
   /**
    * Delete the selected hero
-   * @param {number} id the hero id
    */
   delete(id: number) {
     if (confirm('Are you sure do you want to delete this Game?')) {
