@@ -2,10 +2,10 @@ import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute} from '@angular/router';
 import {AppState} from '../../app.state';
 import {Store} from '@ngrx/store';
-import {GetGame} from '../store/games.actions';
-import {Observable} from 'rxjs/Observable';
-import {Game} from '../shared/game';
 import * as gameActions from '../store/games.actions';
+import {GetGame} from '../store/games.actions';
+import {Observable} from 'rxjs';
+import {Game} from '../shared/game';
 import {getGame} from '../store/games.reducers';
 import {Platform} from '../shared/platform';
 import {getAllPlatforms} from '../store/platforms.reducers';
@@ -26,7 +26,7 @@ export class GameDetailComponent implements OnInit {
 
   ngOnInit() {
     this.route.params.subscribe(params => {
-      this.store.dispatch(new GetGame(+params['id']));
+      this.store.dispatch(new GetGame(+params.id));
     });
     this.platforms = this.store.select(getAllPlatforms);
     this.game = this.store.select(getGame);
@@ -34,7 +34,6 @@ export class GameDetailComponent implements OnInit {
 
   /**
    * Delete the selected hero
-   * @param {number} id the hero id
    */
   delete(id: number) {
     if (confirm('Are you sure do you want to delete this Game?')) {
