@@ -10,11 +10,49 @@ import {Store, StoreModule} from '@ngrx/store';
 import {EffectsModule} from '@ngrx/effects';
 import {PlatformsService} from '../shared/platforms.service';
 import {GamesService} from '../shared/games.service';
-import {MockStore} from '../store/mock-store';
+import * as gamesReducer from '../store/games.reducers';
+import * as platformsReducer from '../store/platforms.reducers';
+import {MockStore, provideMockStore} from '@ngrx/store/testing';
 
 describe('GameEditComponent', () => {
   let component: GameEditComponent;
   let fixture: ComponentFixture<GameEditComponent>;
+  let mockStore: MockStore<{ games: gamesReducer.State, platforms: platformsReducer.State }>;
+  const initialState = {
+    games: {
+      data: [],
+      selected: {
+        id: 1,
+        image: 'horizon_zero_dawn.jpg',
+        name: 'Horizon Zero Dawn',
+        releaseDate: '2017-02-28',
+        platforms: [
+          2
+        ],
+        description: 'Horizon Zero Dawn is an action role-playing video game developed by Guerrilla Games'
+      },
+      action: 'UPDATE_GAME',
+      done: true
+    },
+    platforms: {
+      data: [
+        {
+          id: 1,
+          name: 'Xbox One'
+        },
+        {
+          id: 2,
+          name: 'PlayStation 4'
+        },
+        {
+          id: 3,
+          name: 'PC'
+        }
+      ],
+      action: 'GET_PLATFORMS',
+      done: true
+    }
+  };
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
@@ -35,46 +73,12 @@ describe('GameEditComponent', () => {
         GamesService,
         PlatformsService,
         {provide: APP_BASE_HREF, useValue: '/'},
-        {
-          provide: Store, useValue: new MockStore({
-          games: {
-            data: [],
-            selected: {
-              'id': 1,
-              'image': 'horizon_zero_dawn.jpg',
-              'name': 'Horizon Zero Dawn',
-              'releaseDate': '2017-02-28',
-              'platforms': [
-                2
-              ],
-              'description': 'Horizon Zero Dawn is an action role-playing video game developed by Guerrilla Games'
-            },
-            action: 'UPDATE_GAME',
-            done: true
-          },
-          platforms: {
-            data: [
-              {
-                'id': 1,
-                'name': 'Xbox One'
-              },
-              {
-                'id': 2,
-                'name': 'PlayStation 4'
-              },
-              {
-                'id': 3,
-                'name': 'PC'
-              }
-            ],
-            action: 'GET_PLATFORMS',
-            done: true
-          }
-        })
-        }
+        provideMockStore({initialState})
       ]
     })
       .compileComponents();
+
+    mockStore = TestBed.get(Store);
   }));
 
   beforeEach(() => {
