@@ -1,7 +1,7 @@
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
 import {GameListComponent} from './game-list.component';
-import {BrowserModule, By} from '@angular/platform-browser';
+import {BrowserModule} from '@angular/platform-browser';
 import {APP_BASE_HREF, CommonModule} from '@angular/common';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {HttpClientModule} from '@angular/common/http';
@@ -11,11 +11,62 @@ import {EffectsModule} from '@ngrx/effects';
 import {GamesService} from '../shared/games.service';
 import {PlatformsService} from '../shared/platforms.service';
 import {ExtractNamesPipe} from '../../shared/extract-names.pipe';
-import {MockStore} from '../store/mock-store';
+import * as gamesReducer from '../store/games.reducers';
+import * as platformsReducer from '../store/platforms.reducers';
+import {MockStore, provideMockStore} from '@ngrx/store/testing';
 
 describe('GameListComponent', () => {
   let component: GameListComponent;
   let fixture: ComponentFixture<GameListComponent>;
+  let mockStore: MockStore<{ games: gamesReducer.State, platforms: platformsReducer.State }>;
+  const initialState = {
+    games: {
+      data: [
+        {
+          id: 1,
+          image: 'horizon_zero_dawn.jpg',
+          name: 'Horizon Zero Dawn',
+          releaseDate: '2017-02-28',
+          platforms: [
+            2
+          ],
+          description: 'Horizon Zero Dawn is an action role-playing video game developed by Guerrilla'
+        }, {
+          id: 2,
+          image: 'destiny2.jpg',
+          name: 'Destiny 2',
+          releaseDate: '2017-09-06',
+          platforms: [
+            1,
+            2,
+            3
+          ],
+          description: 'Destiny 2 is an online-only multiplayer first-person shooter video game developed by Bungie'
+        }
+      ],
+      selected: null,
+      action: 'GET_GAMES',
+      done: true
+    },
+    platforms: {
+      data: [
+        {
+          id: 1,
+          name: 'Xbox One'
+        },
+        {
+          id: 2,
+          name: 'PlayStation 4'
+        },
+        {
+          id: 3,
+          name: 'PC'
+        }
+      ],
+      action: 'GET_PLATFORMS',
+      done: true
+    }
+  };
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
@@ -37,59 +88,12 @@ describe('GameListComponent', () => {
         GamesService,
         PlatformsService,
         {provide: APP_BASE_HREF, useValue: '/'},
-        {
-          provide: Store, useValue: new MockStore({
-          games: {
-            data: [
-              {
-                'id': 1,
-                'image': 'horizon_zero_dawn.jpg',
-                'name': 'Horizon Zero Dawn',
-                'releaseDate': '2017-02-28',
-                'platforms': [
-                  2
-                ],
-                'description': 'Horizon Zero Dawn is an action role-playing video game developed by Guerrilla'
-              }, {
-                'id': 2,
-                'image': 'destiny2.jpg',
-                'name': 'Destiny 2',
-                'releaseDate': '2017-09-06',
-                'platforms': [
-                  1,
-                  2,
-                  3
-                ],
-                'description': 'Destiny 2 is an online-only multiplayer first-person shooter video game developed by Bungie'
-              }
-            ],
-            selected: null,
-            action: 'GET_GAMES',
-            done: true
-          },
-          platforms: {
-            data: [
-              {
-                'id': 1,
-                'name': 'Xbox One'
-              },
-              {
-                'id': 2,
-                'name': 'PlayStation 4'
-              },
-              {
-                'id': 3,
-                'name': 'PC'
-              }
-            ],
-            action: 'GET_PLATFORMS',
-            done: true
-          }
-        })
-        }
+        provideMockStore({initialState})
       ]
     })
       .compileComponents();
+
+    mockStore = TestBed.get(Store);
   }));
 
   beforeEach(() => {
